LayoutTests/fast/dom/Window/element-constructors-on-window.html

<html>
<body>
<p>Test that HTML element constructors are accessible as window object properties.</p>
<script>
if (window.layoutTestController)
  layoutTestController.dumpAsText();

function test(className) {
  document.write("<p>" + className + ": " + eval("window." + className) + "</p>");
}

test("HTMLElement");

// DOM 2 elements
test("HTMLAnchorElement");
test("HTMLAppletElement");
test("HTMLAreaElement");
test("HTMLBaseElement");
test("HTMLBaseFontElement");
test("HTMLBodyElement");
test("HTMLBRElement");
test("HTMLButtonElement");
test("HTMLDirectoryElement");
test("HTMLDivElement");
test("HTMLDListElement");
test("HTMLFieldSetElement");
test("HTMLFontElement");
test("HTMLFormElement");
test("HTMLFrameElement");
test("HTMLFrameSetElement");
test("HTMLHeadElement");
test("HTMLHeadingElement");
test("HTMLHRElement");
test("HTMLHtmlElement");
test("HTMLIFrameElement");
test("HTMLImageElement");
test("HTMLInputElement");
test("HTMLLabelElement");
test("HTMLLegendElement");
test("HTMLLIElement");
test("HTMLLinkElement");
test("HTMLMapElement");
test("HTMLMenuElement");
test("HTMLMetaElement");
test("HTMLModElement");
test("HTMLObjectElement");
test("HTMLOListElement");
test("HTMLOptGroupElement");
test("HTMLOptionElement");
test("HTMLParagraphElement");
test("HTMLParamElement");
test("HTMLPreElement");
test("HTMLQuoteElement");
test("HTMLScriptElement");
test("HTMLSelectElement");
test("HTMLStyleElement");
test("HTMLTableCaptionElement");
test("HTMLTableCellElement");
test("HTMLTableColElement");
test("HTMLTableElement");
test("HTMLTableRowElement");
test("HTMLTableSectionElement");
test("HTMLTextAreaElement");
test("HTMLTitleElement");
test("HTMLUListElement");

// Some other elements
test("HTMLCanvasElement");
test("HTMLInsElement");
test("HTMLDelElement");
test("HTMLEmbedElement");
test("HTMLMarqueeElement");

</script>
</body>
</html>
